// src/interfaces/user-detail.ts
export interface UserDetail {
  approved: boolean;
  userId: number;
  name: string;
  surname: string;
  username: string;
  email: string;
  telefono: string;
  codiceFiscale: string;
  consensoPrivacy: boolean;
  terminiCondizioni: boolean;
  marketing: boolean;
  role: string;
  comune: string;
}