// src/interfaces/register.interface.ts
export interface Register {
    userId?: number;
    name: string;
    surname: string;
    username: string;
    email: string;
    password?: string;
    telefono: string;
    codiceFiscale: string;
    consensoPrivacy: boolean;
    terminiCondizioni: boolean;
    marketing: boolean;
    role: string;
    comune: string;
  }
  